Initial checkin of cutting stock program.
This already is complete enough to solve my problem: given a variety of
shapes with a distinct grain bias, how can I best fit them into a board of
a given breadth (in this program called height) to minimize the length of
board that is needed (called width).  The program takes for input any
number of greyscale binary PPM files (P5), a board height, and a gap (for a
cutting margin between pieces, but not at the edges of the board).  After
processing, an optimal length-minimizing solution is given as a ppm file
showing the layout of the pieces and also the final length is printed.

The program can run in multiprocessing mode to utilize all cores or cpus
of a computer.  I still need to document it more thoroughly and do some
other cosmetics.  I can trivially imagine additional features, but this is
sufficient for my purposes.